Move Driver and Server modules underneath the phantom module.
